python-challenge Columbia Module Challenge 3
In the making of both PyBank and PyPoll,
csvreaders were used to read the dataset inside the included .csv's
sys.stdout (import sys) was used to output to create, write and output to textfiles results that
were required for this challenge.

Credits to https://github.com/zunicd/Vote-Counting/blob/master/main.py
which I found while looking for ways to manipulate dictionaries, discovering methods for it such as sum() and max().

All-in-all, had fun!